Moves count actions and reducer into own folder
