+++ src/ducks/counter/counterActions.js
@@ -1,0 +1,4 @@
+export const INCREMENT_REQUESTED = 'increment-requested';
+export const INCREMENT = 'increment';
+export const increment = () => ({ type: INCREMENT });
+export const requestIncrement = () => ({ type: INCREMENT_REQUESTED });
